refactor : ChoiceChips를 signup 공용 컴포넌트로 승격
가입 본인 확인 스텝(#173)의 성별 선택에 재사용하도록
verification 세그먼트에서 signup/_shared/ui로 이동
